xong #63

+++ src/services/OrderService.js
@@ -2,9 +2,8 @@ const User = require("../models/UserModel");
 const bcrypt = require("bcrypt");
 const { generalAccessToken, generalRefreshToken } = require("./JwtService");
 const { JsonWebTokenError } = require("jsonwebtoken");
-const Order = require("../models/OrderProduct")
-const Product = require("../models/ProductModel")
-
+const Order = require("../models/OrderProduct");
+const Product = require("../models/ProductModel");
 
 const createOrder = (newOrder) => {
   return new Promise(async (resolve, reject) => {
@@ -65,17 +64,17 @@ const createOrder = (newOrder) => {
         }
       });
       const results = await Promise.all(promises);
-      const newData = results && results.filter((item) => item.id)
-      if(newData.length){
+      const newData = results && results.filter((item) => item.id);
+      if (newData.length) {
         resolve({
           status: "ERR",
-          message:  `Sản phẩm với id ${newData.join(",")} không đủ hàng`
-        })
+          message: `Sản phẩm với id ${newData.join(",")} không đủ hàng`,
+        });
       }
       resolve({
         status: "OK",
-        message: "success"
-      })
+        message: "success",
+      });
     } catch (e) {
       reject(e);
     }
@@ -85,7 +84,7 @@ const getAllOrderDetails = (id) => {
   return new Promise(async (resolve, reject) => {
     try {
       const order = await Order.find({
-        user: id
+        user: id,
       });
       // console.log("checkUser", checkUser);
       if (order === null) {
@@ -109,7 +108,7 @@ const getOrderDetails = (id) => {
   return new Promise(async (resolve, reject) => {
     try {
       const order = await Order.findById({
-        _id: id
+        _id: id,
       });
       // console.log("checkUser", checkUser);
       if (order === null) {
@@ -129,21 +128,46 @@ const getOrderDetails = (id) => {
     }
   });
 };
-const cancelOrderDetails = (id) => {
+const cancelOrderDetails = (id, data) => {
   return new Promise(async (resolve, reject) => {
     try {
-      const order = await Order.findByIdAndDelete(id);
-      // console.log("checkUser", checkUser);
-      if (order === null) {
-        resolve({
-          status: "OK",
-          message: "The order is not exist!",
-        });
-      }
+      let order = [];
 
+      const promises = data.map(async (order) => {
+        const productData = await Product.findOneAndUpdate(
+          {
+            _id: order.product,
+            sold: {$gte: order.amount}
+          },
+          {
+            $inc: {
+              countInStock: +order.amount,
+              sold: -order.amount,
+            },
+          },
+          { new: true }
+        );
+        if (productData) {
+          order = await Order.findByIdAndDelete(id);
+          if (order === null) {
+            resolve({
+              status: "ERR",
+              message: "The order is not define",
+            });
+          }
+        } else {
+          return {
+            status: "OK",
+            message: "ERR",
+            id: order.product,
+          };
+        }
+      });
+      const results = await Promise.all(promises);
+     
       resolve({
         status: "OK",
-        message: "SUCCESS!",
+        message: "success",
         data: order,
       });
     } catch (e) {
@@ -151,11 +175,10 @@ const cancelOrderDetails = (id) => {
     }
   });
 };
-
 
 module.exports = {
   createOrder,
   getAllOrderDetails,
   getOrderDetails,
-  cancelOrderDetails
+  cancelOrderDetails,
 };
